backend/requirements.txt
# Core API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0

# Database (опционально, если используется pgvector)
asyncpg>=0.29.0
psycopg[binary]>=3.1.0

# Vector DB & Embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
transformers>=4.35.0,<4.39

# LLM clients
openai>=1.3.0
google-generativeai>=0.3.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0

# MCP (Model Context Protocol)
mcp>=1.11.0
langchain-mcp-adapters>=0.1.9

# LangGraph для агентов
langgraph>=0.2.0

# Web UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Image processing (для Vision Agent)
Pillow>=10.0.0

# Web parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.25.0

# PDF parsing
PyPDF2>=3.0.0

# OpenCLIP для мультимодальных эмбеддингов (из sam2seg)
open-clip-torch>=2.24.0
torch>=2.0.0
torchvision>=0.15.0

# Logging
python-json-logger>=2.0.7

